+++ src/include/Cell.hpp
@@ -5,7 +5,7 @@
 class Cell
 {
 public:
-	Cell(int x, int y, int ID, int colum, int CellMS, Color c);
+	Cell(int x, int y, int ID, int colum, float CellMS, Color c);
 
 	void Draw();
 	void Update(const Vector2& mousePos);
@@ -15,7 +15,7 @@ public:
 	int ReturnID() const;
 	bool GetActive() const;
 	Vector2 GetPosition() const;
-	int GetCellMS() const;
+	float GetCellMS() const;
 	int GetID() const;
 
 	Vector2 GetNextCellPosition(int direction) const;
@@ -30,5 +30,5 @@ private:
 	bool m_Active;
 
 	int m_ID;
-	int m_CellMS;
+	float m_CellMS;
 };
